refactor(heureka): remove unnecessary useEffect
